school_portal/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.2.1
cohere==5.20.0
colorama==0.4.6
debugpy==1.8.17
dj-database-url==3.0.1
Django==5.2.7
django-widget-tweaks==1.5.0
fastavro==1.12.1
filelock==3.20.0
fsspec==2025.10.0
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==1.0.1
idna==3.11
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
packaging==25.0
parso==0.8.5
pillow==12.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.3
tokenizers==0.22.1
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
django-environ==0.11.2
black==24.3.0

whitenoise==6.6.0
mypy==1.9.0
